feat(vue): rebuild the project based on the latest starter
