+++ android/app/src/main/res/values/strings.xml
@@ -1,7 +1,7 @@
 <?xml version='1.0' encoding='utf-8'?>
 <resources>
-    <string name="app_name">Vue Photo Gallery</string>
-    <string name="title_activity_main">Vue Photo Gallery</string>
-    <string name="package_name">io.ionic.demo.pg.vue</string>
-    <string name="custom_url_scheme">io.ionic.demo.pg.vue</string>
+    <string name="app_name">photo-gallery</string>
+    <string name="title_activity_main">photo-gallery</string>
+    <string name="package_name">io.ionic.starter</string>
+    <string name="custom_url_scheme">io.ionic.starter</string>
 </resources>
