+++ android/variables.gradle
@@ -7,10 +7,10 @@ ext {
     androidxCoordinatorLayoutVersion = '1.2.0'
     androidxCoreVersion = '1.8.0'
     androidxFragmentVersion = '1.4.1'
+    coreSplashScreenVersion = '1.0.0-rc01'
+    androidxWebkitVersion = '1.4.0'
     junitVersion = '4.13.2'
     androidxJunitVersion = '1.1.3'
     androidxEspressoCoreVersion = '3.4.0'
     cordovaAndroidVersion = '10.1.1'
-    coreSplashScreenVersion = '1.0.0-rc01'
-    androidxWebkitVersion = '1.4.0'
 }
