+++ capacitor.config.ts
@@ -1,15 +1,15 @@
-import { CapacitorConfig} from "@capacitor/cli";
+import { CapacitorConfig } from '@capacitor/cli';
+
 const config: CapacitorConfig = {
-  appId: "io.ionic.demo.pg.vue",
-  appName: "Vue Photo Gallery",
+  appId: 'io.ionic.demo.pg.vue',
+  appName: 'Vue Photo Gallery',
   bundledWebRuntime: false,
-  npmClient: "npm",
-  webDir: "dist",
+  webDir: 'dist',
   plugins: {
     SplashScreen: {
-      launchShowDuration: 0
-    }
-  }
-}
+      launchShowDuration: 0,
+    },
+  },
+};
 
-export default config
+export default config;
