+++ package.json
@@ -7,41 +7,55 @@
     "build": "vue-cli-service build",
     "test:unit": "vue-cli-service test:unit",
     "test:e2e": "vue-cli-service test:e2e",
-    "lint": "vue-cli-service lint"
+    "lint": "vue-cli-service lint",
+    "prepare": "husky install"
   },
   "dependencies": {
-    "@capacitor/android": "^4.0.0",
-    "@capacitor/camera": "^4.0.0",
-    "@capacitor/core": "^4.0.0",
-    "@capacitor/filesystem": "^4.0.0",
-    "@capacitor/ios": "^4.0.0",
-    "@capacitor/preferences": "^4.0.1",
-    "@ionic/pwa-elements": "^3.0.1",
-    "@ionic/vue": "^5.6.7",
-    "@ionic/vue-router": "^5.6.7",
+    "@capacitor/android": "4.6.1",
+    "@capacitor/app": "4.1.1",
+    "@capacitor/camera": "~4.1.4",
+    "@capacitor/core": "4.6.1",
+    "@capacitor/filesystem": "~4.1.4",
+    "@capacitor/haptics": "4.1.0",
+    "@capacitor/ios": "4.6.1",
+    "@capacitor/keyboard": "4.1.0",
+    "@capacitor/preferences": "~4.0.2",
+    "@capacitor/status-bar": "4.1.1",
+    "@ionic/pwa-elements": "~3.1.1",
+    "@ionic/vue": "^6.0.0",
+    "@ionic/vue-router": "^6.0.0",
     "core-js": "^3.6.5",
-    "vue": "^3.0.11",
-    "vue-router": "^4.0.8"
+    "ionicons": "^6.0.3",
+    "vue": "^3.2.21",
+    "vue-router": "^4.0.12"
   },
   "devDependencies": {
-    "@capacitor/cli": "^4.0.0",
-    "@types/jest": "^24.0.19",
-    "@typescript-eslint/eslint-plugin": "^2.33.0",
-    "@typescript-eslint/parser": "^2.33.0",
-    "@vue/cli-plugin-babel": "~4.5.0",
-    "@vue/cli-plugin-e2e-cypress": "~4.5.0",
-    "@vue/cli-plugin-eslint": "~4.5.0",
-    "@vue/cli-plugin-router": "~4.5.0",
-    "@vue/cli-plugin-typescript": "~4.5.0",
-    "@vue/cli-plugin-unit-jest": "~4.5.0",
-    "@vue/cli-service": "4.5.6",
-    "@vue/compiler-sfc": "^3.0.0-0",
-    "@vue/eslint-config-typescript": "^5.0.2",
-    "@vue/test-utils": "^2.0.0-0",
-    "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^7.0.0-0",
-    "typescript": "~3.9.3",
-    "vue-jest": "^5.0.0-0"
+    "@capacitor/cli": "4.6.1",
+    "@ionic/prettier-config": "~2.0.0",
+    "@types/jest": "^27.0.2",
+    "@typescript-eslint/eslint-plugin": "^5.6.0",
+    "@typescript-eslint/parser": "^5.6.0",
+    "@vue/cli-plugin-babel": "~5.0.0-rc.1",
+    "@vue/cli-plugin-e2e-cypress": "~5.0.0-rc.1",
+    "@vue/cli-plugin-eslint": "~5.0.0-rc.1",
+    "@vue/cli-plugin-router": "~5.0.0-rc.1",
+    "@vue/cli-plugin-typescript": "~5.0.0-rc.1",
+    "@vue/cli-plugin-unit-jest": "~5.0.0-rc.1",
+    "@vue/cli-service": "~5.0.0-rc.1",
+    "@vue/eslint-config-typescript": "^11.0.2",
+    "@vue/test-utils": "^2.0.0-rc.16",
+    "@vue/vue3-jest": "^27.0.0-alpha.3",
+    "babel-jest": "^27.3.1",
+    "cypress": "^8.7.0",
+    "eslint": "^8.4.1",
+    "eslint-plugin-vue": "^9.8.0",
+    "husky": "~8.0.3",
+    "jest": "^27.3.1",
+    "prettier": "~2.8.1",
+    "pretty-quick": "~3.1.3",
+    "ts-jest": "^27.0.7",
+    "typescript": "^4.3.5"
   },
+  "prettier": "@ionic/prettier-config",
   "description": "Build a photo gallery app that runs on iOS, Android, and the web - with just one codebase."
 }
