+++ src/composables/usePhotoGallery.ts
@@ -1,19 +1,35 @@
 import { ref, onMounted, watch } from 'vue';
-import { Capacitor } from '@capacitor/core';
-import {
-  Camera,
-  CameraSource,
-  CameraResultType,
-  Photo,
-} from '@capacitor/camera';
+import { Camera, CameraResultType, CameraSource, Photo } from '@capacitor/camera';
 import { Filesystem, Directory } from '@capacitor/filesystem';
 import { Preferences } from '@capacitor/preferences';
-
 import { isPlatform } from '@ionic/vue';
+import { Capacitor } from '@capacitor/core';
 
-export function usePhotoGallery() {
-  const photos = ref<UserPhoto[]>([]);
+export interface UserPhoto {
+  filepath: string;
+  webviewPath?: string;
+}
+
+export const usePhotoGallery = () => {
   const PHOTO_STORAGE = 'photos';
+  const photos = ref<UserPhoto[]>([]);
+
+  const convertBlobToBase64 = (blob: Blob): Promise<string> =>
+    new Promise((resolve, reject) => {
+      const reader = new FileReader();
+      reader.onerror = reject;
+      reader.onload = () => {
+        resolve(reader.result as string);
+      };
+      reader.readAsDataURL(blob);
+    });
+
+  const cachePhotos = () => {
+    Preferences.set({
+      key: PHOTO_STORAGE,
+      value: JSON.stringify(photos.value),
+    });
+  };
 
   const loadSaved = async () => {
     const photoList = await Preferences.get({ key: PHOTO_STORAGE });
@@ -34,31 +50,16 @@ export function usePhotoGallery() {
     photos.value = photosInPreferences;
   };
 
-  const convertBlobToBase64 = (blob: Blob) =>
-    new Promise((resolve, reject) => {
-      const reader = new FileReader();
-      reader.onerror = reject;
-      reader.onload = () => {
-        resolve(reader.result);
-      };
-      reader.readAsDataURL(blob);
-    });
-
-  const savePicture = async (
-    photo: Photo,
-    fileName: string
-  ): Promise<UserPhoto> => {
+  const savePicture = async (photo: Photo, fileName: string): Promise<UserPhoto> => {
     let base64Data: string;
-    // "hybrid" will detect Cordova or Capacitor;
+    // "hybrid" will detect mobile - iOS or Android
     if (isPlatform('hybrid')) {
       const file = await Filesystem.readFile({
-        // eslint-disable-next-line
         path: photo.path!,
       });
       base64Data = file.data;
     } else {
       // Fetch the photo, read as a blob, then convert to base64 format
-      // eslint-disable-next-line
       const response = await fetch(photo.webPath!);
       const blob = await response.blob();
       base64Data = (await convertBlobToBase64(blob)) as string;
@@ -86,18 +87,6 @@ export function usePhotoGallery() {
     }
   };
 
-  const takePhoto = async () => {
-    const photo = await Camera.getPhoto({
-      resultType: CameraResultType.Uri,
-      source: CameraSource.Camera,
-      quality: 100,
-    });
-    const fileName = new Date().getTime() + '.jpeg';
-    const savedFileImage = await savePicture(photo, fileName);
-
-    photos.value = [savedFileImage, ...photos.value];
-  };
-
   const deletePhoto = async (photo: UserPhoto) => {
     // Remove this photo from the Photos reference data array
     photos.value = photos.value.filter((p) => p.filepath !== photo.filepath);
@@ -110,25 +99,25 @@ export function usePhotoGallery() {
     });
   };
 
-  const cachePhotos = () => {
-    Preferences.set({
-      key: PHOTO_STORAGE,
-      value: JSON.stringify(photos.value),
+  const takePhoto = async () => {
+    const photo = await Camera.getPhoto({
+      resultType: CameraResultType.Uri,
+      source: CameraSource.Camera,
+      quality: 100,
     });
+
+    const fileName = new Date().getTime() + '.jpeg';
+    const savedFileImage = await savePicture(photo, fileName);
+
+    photos.value = [savedFileImage, ...photos.value];
   };
 
   onMounted(loadSaved);
-
   watch(photos, cachePhotos);
 
   return {
     photos,
-    takePhoto,
     deletePhoto,
+    takePhoto,
   };
-}
-
-export interface UserPhoto {
-  filepath: string;
-  webviewPath?: string;
-}
+};
