+++ src/main.ts
@@ -1,5 +1,5 @@
-import { createApp } from 'vue'
-import App from './App.vue'
+import { createApp } from 'vue';
+import App from './App.vue';
 import router from './router';
 
 import { IonicVue } from '@ionic/vue';
@@ -19,17 +19,15 @@ import '@ionic/vue/css/text-alignment.css';
 import '@ionic/vue/css/text-transformation.css';
 import '@ionic/vue/css/flex-utils.css';
 import '@ionic/vue/css/display.css';
+import { defineCustomElements } from '@ionic/pwa-elements/loader';
 
 /* Theme variables */
 import './theme/variables.css';
 
-import { defineCustomElements } from '@ionic/pwa-elements/loader';
 defineCustomElements(window);
 
-const app = createApp(App)
-  .use(IonicVue)
-  .use(router);
-  
+const app = createApp(App).use(IonicVue).use(router);
+
 router.isReady().then(() => {
   app.mount('#app');
 });
