+++ tests/e2e/.eslintrc.js
@@ -1,12 +1,10 @@
 module.exports = {
-  plugins: [
-    'cypress'
-  ],
+  plugins: ['cypress'],
   env: {
     mocha: true,
-    'cypress/globals': true
+    'cypress/globals': true,
   },
   rules: {
-    strict: 'off'
-  }
-}
+    strict: 'off',
+  },
+};
