+++ tests/e2e/plugins/index.js
@@ -20,6 +20,6 @@ module.exports = (on, config) => {
     integrationFolder: 'tests/e2e/specs',
     screenshotsFolder: 'tests/e2e/screenshots',
     videosFolder: 'tests/e2e/videos',
-    supportFile: 'tests/e2e/support/index.js'
-  })
-}
+    supportFile: 'tests/e2e/support/index.js',
+  });
+};
