+++ tests/unit/example.spec.ts
@@ -1,12 +1,9 @@
-import { shallowMount } from '@vue/test-utils'
-import HelloWorld from '@/components/HelloWorld.vue'
+import { mount } from '@vue/test-utils';
+import Tab1Page from '@/views/Tab1Page.vue';
 
-describe('HelloWorld.vue', () => {
-  it('renders props.msg when passed', () => {
-    const msg = 'new message'
-    const wrapper = shallowMount(HelloWorld, {
-      props: { msg }
-    })
-    expect(wrapper.text()).toMatch(msg)
-  })
-})
+describe('Tab1Page.vue', () => {
+  it('renders tab 1 Tab1Page', () => {
+    const wrapper = mount(Tab1Page);
+    expect(wrapper.text()).toMatch('Tab 1 page');
+  });
+});
